Finished some TODOs, minor refactorings - replaced some TODOs with information - converted some string properties to use enums - minor property name/capitolization refactoring - fixed some xml doc comments - GetPoint is now ToPoint - added ToLocation - changes AL.Pathfinding usage of map key to use accessor instead

+++ AL.Client/Abstractions/AsyncDeltaLoop.cs
@@ -73,8 +73,8 @@ namespace AL.Client.Abstractions
                     Client.Logger.Error(ex);
                 } finally
                 {
-                    var executionTime = DeltaTime.Value - delta;
-                    var timeTillNextLoop = Convert.ToInt32(1000 / PollingRate - executionTime);
+                    var executionTime = (int)(DeltaTime.Value - delta);
+                    var timeTillNextLoop = Convert.ToInt32(1000 / PollingRate) - executionTime;
 
                     if (timeTillNextLoop > 0)
                         await Task.Delay(timeTillNextLoop).ConfigureAwait(false);
